Move parser implementations & tests under compare_locales/parsers/
