+++ compare_locales/parsers/base.py
