+++ compare_locales/parsers/defines.py
