+++ compare_locales/parsers/properties.py
