+++ compare_locales/parsers/tests/data/bug121341.properties
