+++ compare_locales/parsers/tests/data/test.properties
