+++ compare_locales/parsers/tests/data/triple-license.dtd
