+++ compare_locales/parsers/tests/test_android.py
