+++ compare_locales/parsers/tests/test_properties.py
